requirements.txt
scapy>=2.5.0
python-nmap>=0.7.1
netfilterqueue>=0.8.1
pycryptodome>=3.19.0
hashcat>=0.0.1
tkinter
matplotlib>=3.7.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tensorflow>=2.13.0
pcapy>=0.11.5
dpkt>=1.9.8
colorama>=0.4.6
pyfiglet>=1.0.2


